add UI X-ray widget
