+++ web/package.json
@@ -42,6 +42,7 @@
     "react-redux": "^5.0.6",
     "react-router-dom": "^4.2.2",
     "react-router-redux": "^5.0.0-alpha.9",
+    "react-x-ray": "^1.0.0-4",
     "rebass": "^1.0.4",
     "redux": "^3.7.2",
     "styled-components": "^2.3.3"
